actv1/requirements.txt
fastapi==0.115.12
uvicorn[standard]==0.34.0
sqlalchemy==2.0.41
alembic==1.16.0
psycopg[binary]==3.2.9
redis==6.0.0
pydantic-settings==2.9.1
passlib==1.7.4
python-jose[cryptography]==3.4.0
python-multipart==0.0.20
xgboost==2.1.4
scikit-learn==1.5.2
networkx==3.3
PuLP==2.9.0
aiosqlite==0.20.0
requests==2.32.3
